retrieve/requirements.txt
elasticsearch==8.3.1
beautifulsoup4
hazm
numpy
networkx
pandas
torch
scikit-learn
matplotlib
transformers
datasets
gensim
tqdm
grpcio
protobuf